refactor: assign Globe to collect vendir secrets before sync
